fixing some timestamp type difference warnings...
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@12669 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/clif.c
@@ -11414,7 +11414,7 @@ void clif_Mail_refreshinbox(struct map_session_data *sd)
 		memcpy(WFIFOP(fd,12+73*j), msg->title, MAIL_TITLE_LENGTH);
 		WFIFOB(fd,52+73*j) = (msg->status != MAIL_UNREAD); // 0: unread, 1: read
 		memcpy(WFIFOP(fd,53+73*j), msg->send_name, NAME_LENGTH);
-		WFIFOL(fd,77+73*j) = msg->timestamp;
+		WFIFOL(fd,77+73*j) = (uint32)msg->timestamp;
 		j++;
 	}
 	WFIFOSET(fd,len);
@@ -11763,7 +11763,7 @@ void clif_Auction_results(struct map_session_data *sd, short count, short pages,
 		WFIFOL(fd,47+k) = auction.price;
 		WFIFOL(fd,51+k) = auction.buynow;
 		safestrncpy((char*)WFIFOP(fd,55+k), auction.buyer_name, NAME_LENGTH);
-		WFIFOL(fd,79+k) = auction.timestamp;
+		WFIFOL(fd,79+k) = (uint32)auction.timestamp;
 	}
 	WFIFOSET(fd, 12 + (count * 83));
 }
